+++ Packages/OpenApiCodeGen/Editor/Domain/Progress/FailedProgressStatus.cs
@@ -1,0 +1,17 @@
+﻿#nullable enable
+namespace ReBeat.OpenApiCodeGen.UI
+{
+    internal class FailedProgressStatus : IProgressStatus
+    {
+        public double Progress => 0.0;
+        /// <summary>
+        /// 失敗理由
+        /// </summary>
+        public string? Reason { get; }
+
+        public FailedProgressStatus(string? reason = null)
+        {
+            Reason = reason;
+        }
+    }
+}