+++ Packages/OpenApiCodeGen/Editor/Service/Dto/CheckDockerInstalledDto.cs
@@ -1,0 +1,13 @@
+﻿namespace ReBeat.OpenApiCodeGen.Core
+{
+    internal class CheckDockerInstalledDto
+    {
+        public string DockerPath { get; private set; }
+
+        public CheckDockerInstalledDto(string dockerPath)
+        {
+            this.DockerPath = dockerPath;
+        }
+    }
+
+}