+++ Packages/OpenApiCodeGen/Editor/Service/Dto/SetupDto.cs
@@ -1,0 +1,21 @@
+﻿
+namespace ReBeat.OpenApiCodeGen.Core
+{
+    internal class SetupDto
+    {
+        public string DockerPath { get; }
+        public GenerateProvider ProviderType { get; }
+
+        public SetupDto()
+        {
+            DockerPath = string.Empty;
+            ProviderType = GenerateProvider.OpenApi;
+        }
+
+        public SetupDto(string dockerPath, GenerateProvider providerType)
+        {
+            DockerPath = dockerPath;
+            ProviderType = providerType;
+        }
+    }
+}