+++ Packages/OpenApiCodeGen/Editor/Share/GenerateProvider.cs
@@ -1,0 +1,8 @@
+﻿namespace ReBeat.OpenApiCodeGen.Core
+{
+    internal enum GenerateProvider
+    {
+        OpenApi,
+    }
+
+}